results and table display on separate pages for each selected team
